% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logl.R
\name{tidy_logl}
\alias{tidy_logl}
\title{Compute log-likelihood values for text mining.}
\usage{
tidy_logl(df, group_key, token_key, threshold = 3)
}
\arguments{
\item{df}{A data frame (or tibble, etc.) in tidytext format}

\item{group_key}{Column name of \code{df} to use for grouping of
log-likelihood}

\item{token_key}{Column name of \code{df} that contains tokens or n-grams}

\item{threshold}{Minimum log-likelihood value (number) to include in results}
}
\value{
A data frame with columns for token, group, and ll (log-likelihood
values above \code{threshold})
}
\description{
Compute log-likelihood values for text mining.
}
